requirements.txt
aiogram==3.0.0b7
requests==2.28.1
python-dotenv==1.0.0

# Web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# AI and ML dependencies
scikit-learn==1.3.2
numpy==1.24.3

# Зависимости для тестирования
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-mock==3.10.0
aresponses==3.0.0
pytest-cov==4.1.0

python-docx==0.8.11
